Setup generic debug map drawing

+++ assets.h
@@ -1,1 +1,1 @@
-int genNeonColor();
+uint16_t genNeonColor();
