+++ color.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 // https://gist.github.com/kuathadianto/200148f53616cbd226d993b400214a7fs
-int HSVtoRGB(int H, double S, double V) {
+uint16_t HSVtoRGB(int H, double S, double V) {
 	double C = S * V;
 	double X = C * (1 - abs(fmod(H / 60.0, 2) - 1));
 	double m = V - C;
@@ -43,7 +43,7 @@ int HSVtoRGB(int H, double S, double V) {
 		Bs = X;
 	}
 
-	int rgb = 0;
+	uint16_t rgb = 0;
 
 	rgb += (Rs + m) * 255 * 65536;  // 65536 = 16 ^ 4 is hex offset to move 0xAA to 0xRR0000
 	rgb += (Gs + m) * 255 * 256; // 256 = 16 ^ 2 is hex offset to move 0xGG to 0x00GG00
@@ -51,6 +51,6 @@ int HSVtoRGB(int H, double S, double V) {
     return rgb;
 }
 
-int genNeonColor() {
+uint16_t genNeonColor() {
 	return HSVtoRGB(random(256), 1.0, 0.05);
 }
